+++ misc-utils/copyfilerange.c
@@ -133,23 +133,24 @@ static void copy_range(struct rangeitem *range) {
 	uintmax_t remaining = range->length;
 
 	if (range->in_offset > range->in_st_size)
-		errx(EXIT_FAILURE, _("%s offset %"PRId64" is beyond file size of %"PRId64""),
-		                     range->in_filename, range->in_offset, range->in_st_size);
+		errx(EXIT_FAILURE, _("%s offset %jd is beyond file size of %jd"),
+		                     range->in_filename, (intmax_t) range->in_offset, (intmax_t) range->in_st_size);
 
 	while (remaining > 0) {
 		const size_t chunk = remaining > SIZE_MAX ? SIZE_MAX : remaining;
 		if (verbose)
-			printf("copy_file_range %s to %s %"PRId64":%"PRId64":%zu\n",
+			printf("copy_file_range %s to %s %jd:%jd:%zu\n",
 			       range->in_filename, range->out_filename,
-			       range->in_offset, range->out_offset, chunk);
+			       (intmax_t) range->in_offset, (intmax_t) range->out_offset, chunk);
 
 		const ssize_t copied = copy_file_range(range->in_fd, &range->in_offset,
 		                                       range->out_fd, &range->out_offset, chunk, 0);
 		if (copied < 0)
-			errx(EXIT_FAILURE, _("failed to copy range %"PRId64":%"PRId64":%ju "
+			errx(EXIT_FAILURE, _("failed to copy range %jd:%jd:%ju "
 			                     "from %s to %s with %ju remaining: %m\n"),
-			                     range->in_offset, range->out_offset, range->length,
-			                     range->in_filename, range->out_filename, remaining);
+			                     (intmax_t) range->in_offset, (intmax_t) range->out_offset,
+								 range->length, range->in_filename,
+								 range->out_filename, remaining);
 		if (copied == 0)
 			break;
